refactoring: tests moved to test15 package and UI handlers hacks moved into TestSampleRoboApplication_
