+++ AndroidAnnotations/functional-test-1-5/src/test/java/org/androidannotations/test15/UiThreadExecutorTest.java
@@ -1,13 +1,10 @@
-package org.androidannotations.api;
+package org.androidannotations.test15;
 
-import android.os.Handler;
-import android.os.Message;
-import org.junit.BeforeClass;
+import org.androidannotations.api.UiThreadExecutor;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.robolectric.Robolectric;
 import org.robolectric.RobolectricTestRunner;
-import org.robolectric.shadows.ShadowHandler;
 
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
@@ -17,23 +14,6 @@ import static org.junit.Assert.*;
 
 @RunWith(RobolectricTestRunner.class)
 public class UiThreadExecutorTest {
-
-	//TODO remove this after upgrading robolectric to 3+
-	@BeforeClass
-	public static void hackOldRobolectric() {
-		final Handler handler = UiThreadExecutor.HANDLER;
-		ShadowHandler shadowHandler = Robolectric. shadowOf_(handler);
-		shadowHandler.__constructor__(new Handler.Callback() {
-			@Override
-			public boolean handleMessage(Message msg) {
-				//in the robolectric 2.4 there is a strange code - they call Handler's handleMessage (it do nothing)
-				//instead of dispatch, that actually do the job. This is just a dirty work-around. It should be removed
-				//with newer version of robolectric.
-				handler.dispatchMessage(msg);
-				return true;
-			}
-		});
-	}
 
 	@Test
 	public void oneTaskTest() throws Exception {
